refactor: move warmup into image.ts, use getPages instead of fs walk
Delete separate image-warmup plugin. Warmup now uses getPages() +
getPageImages() from source.ts (images already collected by remark
plugin). Triggered from ready probe alongside search index.
