+++ packages/chronicle/src/server/api/ready.ts
@@ -1,9 +1,11 @@
 import { defineHandler } from 'nitro';
 import { ensureIndex, isSearchReady } from './search';
+import { warmupImageCache } from './image';
 import { LATEST_CONTEXT } from '@/lib/version-source';
 
 export default defineHandler(async () => {
   ensureIndex(LATEST_CONTEXT).catch(e => console.error('[search:index]', e));
+  warmupImageCache().catch(e => console.error('[image-warmup]', e));
 
   if (!isSearchReady()) {
     return new Response(JSON.stringify({ status: 'not_ready', search: false }), {
